Change windowSize parameter of blur shader to int

+++ Engine/Shaders/BlurFast.frag
@@ -5,12 +5,12 @@ in vec2 pass_texCoords;
 
 out vec4 fragColor;
 
-uniform vec2 windowSize;
+uniform ivec2 windowSize;
 uniform sampler2D tex;
 uniform vec2 direction;
 uniform int mipmap;
 
-vec3 blur(int lod, vec2 uv, vec2 resolution, vec2 direction) {
+vec3 blur(int lod, vec2 uv, ivec2 resolution, vec2 direction) {
     vec3 color = vec3(0);
     vec2 off = vec2(1.333333333333) * direction;
     color += textureLod(tex, uv, lod).rgb * 0.29411764705882354;
